security-command-center/cloud_function/requirements.txt
google-cloud-secret-manager
slack_sdk
